add Input_Data class and write output in new mtehod
