+++ execution.py
@@ -3,9 +3,20 @@ from modules import *
 CALIBRATION = False
 PROFILING = False
 
-geometry = [3, 6, 30, 20, 60, 40]
 
 if __name__ == "__main__":
+    
+    # Define worksheets and location references for inputs
+    app, wb, ctrl, odin, loc_dict = Data_Input.main()
+
+    # Iterate through each parameter to parse values and return checks
+    geometry = []
+    for param in ["v", "e", "l", "r", "icd", "phi"]:
+        geometry.append(Data_Input.import_params(ctrl, param, loc_dict)) 
+
+    od, arms = Data_Input.import_od(ctrl, odin)
+
+    # Import geometry...
     od_builder = OD_Eval(od_type=Flow_Type.RANDOM)
     od = od_builder.rand_od_builder(arms=4)
     if PROFILING == True:
@@ -22,3 +33,5 @@ if __name__ == "__main__":
     model = Capacity_Eval(*geometry, circulatory_flow=od_builder.Qc(arm_index=1, od=od))
     arm_capacity = od_builder.Qe_stack(od)[0] / model.compute()
     print(f'RFC: {round(arm_capacity,3)}')
+
+    Data_Input.print_out(app, wb, ctrl, arm_capacity)
