+++ modules/__init__.py
@@ -3,3 +3,4 @@ from modules.profiler import Profiler, Profiles
 from modules.od_flows import OD_Eval, Flow_Type
 from modules.capacity_eval import Capacity_Eval
 from modules.optimiser import Geo_Optimiser, JunctionType
+from modules.read_data import Data_Input
